requirements.txt
Flask==2.3.2
google-cloud-firestore==2.11.0